[FOLD] Populate sfConsensusHash

+++ src/test/app/RCLValidations_test.cpp
@@ -36,8 +36,8 @@ class RCLValidations_test : public beast::unit_test::suite
     {
         testcase("Change validation trusted status");
         PublicKey key = derivePublicKey(KeyType::ed25519, randomSecretKey());
-        auto v = std::make_shared<STValidation>(
-            uint256(), NetClock::time_point(), key, calcNodeID(key), true);
+        auto v = std::make_shared<STValidation>(uint256(), uint256(),
+            NetClock::time_point(), key, calcNodeID(key), true);
 
         BEAST_EXPECT(!v->isTrusted());
         v->setTrusted();
